Conflicts:
	Raven.Studio/Features/Indexes/EditIndexView.xaml

+++ Raven.Client.Lightweight/Document/RavenClientEnlistment.cs
@@ -7,6 +7,7 @@
 using System;
 using System.IO;
 using System.IO.IsolatedStorage;
+using System.Threading;
 using System.Transactions;
 using Raven.Abstractions.Logging;
 
@@ -57,7 +58,7 @@ namespace Raven.Client.Document
 						file.Flush(true);
 					}
 					machineStoreForApplication.MoveFile(name + ".temp", name);
-			}
+				}
 			}
 			catch (Exception e)
 			{
@@ -81,11 +82,8 @@ namespace Raven.Client.Document
 			{
 				session.Commit(PromotableRavenClientEnlistment.GetLocalOrDistributedTransactionId(transaction));
 
-				using (var machineStoreForApplication = IsolatedStorageFile.GetMachineStoreForDomain())
-				{
-					machineStoreForApplication.DeleteFile(TransactionRecoveryInformationFileName);
-			}
-			}
+				DeleteFile();
+				}
 			catch (Exception e)
 			{
 				logger.ErrorException("Could not commit distributed transaction", e);
@@ -106,16 +104,42 @@ namespace Raven.Client.Document
 			{
 				session.Rollback(PromotableRavenClientEnlistment.GetLocalOrDistributedTransactionId(transaction));
 
-				using (var machineStoreForApplication = IsolatedStorageFile.GetMachineStoreForDomain())
-				{
-					machineStoreForApplication.DeleteFile(TransactionRecoveryInformationFileName);
-				}
+				DeleteFile();
 			}
 			catch (Exception e)
 			{
 				logger.ErrorException("Could not rollback distributed transaction", e);
 			}
 			enlistment.Done(); // will happen anyway, tx will be rolled back after timeout
+		}
+
+		private void DeleteFile()
+		{
+				using (var machineStoreForApplication = IsolatedStorageFile.GetMachineStoreForDomain())
+				{
+				// docs says to retry: http://msdn.microsoft.com/en-us/library/system.io.isolatedstorage.isolatedstoragefile.deletefile%28v=vs.95%29.aspx
+				int retries = 10;
+				while(true)
+				{
+					if (machineStoreForApplication.FileExists(TransactionRecoveryInformationFileName) == false)
+						break;
+					try
+					{
+					machineStoreForApplication.DeleteFile(TransactionRecoveryInformationFileName);
+						break;
+				}
+					catch (IsolatedStorageException)
+					{
+						retries -= 1;
+						if(retries > 0 )
+						{
+							Thread.Sleep(100);
+							continue;
+						}
+						throw;
+					}
+			}
+			}
 		}
 
 		/// <summary>
@@ -129,11 +153,8 @@ namespace Raven.Client.Document
 			{
 				session.Rollback(PromotableRavenClientEnlistment.GetLocalOrDistributedTransactionId(transaction));
 
-				using (var machineStoreForApplication = IsolatedStorageFile.GetMachineStoreForDomain())
-				{
-					machineStoreForApplication.DeleteFile(TransactionRecoveryInformationFileName);
+				DeleteFile();
 				}
-			}
 			catch (Exception e)
 			{
 				logger.ErrorException("Could not mark distriubted transaction as in doubt", e);
@@ -159,10 +180,7 @@ namespace Raven.Client.Document
 			{
 				session.Rollback(PromotableRavenClientEnlistment.GetLocalOrDistributedTransactionId(transaction));
 
-				using (var machineStoreForApplication = IsolatedStorageFile.GetMachineStoreForDomain())
-				{
-					machineStoreForApplication.DeleteFile(TransactionRecoveryInformationFileName);
-				}
+				DeleteFile();
 			}
 			catch (Exception e)
 			{
